Implement sweep selection as alternative for the piece menu
When the option -pieceMenu is off, the piece menu will no longer appear
on right-clicks in Edit Position mode. In stead a Pawn of the chosen
color will appear immediately in the clicked square, and vertical
movement of the mouse with the right button down will cycle through
all other pieces defined in the pieceToCharTable of the current variant.
  Setting the side to move is acheived by clicking the correspondig
clock. Clear board can be effected by clicking the clock of the side
that already has the move.