Fix use of game/position file in first match game
When a match was started from the menu, it was not paying attention to
game and position files. (Later games were.) The code that handled this
in the initialization when a match is started through the command line
was made into a routine MatchEvent, which can be called from the menu
proc too.