Implement move type-in for XBoard
By adding an event handler to the main window for key presses, users can
now type moves when the board has focus, like in WinBoard. The typed
character appears as first character in the type-in box that willpop up.
Enter and Escape close the box; finishing the typing of the move enters
it. Like in WinBoard it is also possible to type a move number (to go
there) or (in Edit Position mode) a FEN.