Couple mouse wheel to v-scrolls in file browser
A newly added mouse-wheel event handler was made to call the v-scroll
callback routines with parameters to step one entry. The button-press
handler for the lists themselves was also made to intercept the wheel
events (buttons 4 and 5), to prevent they would select an entry, and
make them scroll in stead as well.