Fix crash on closing dialogs in wrong order
When a menu dialog was closed when it still had an open file-browse
daughter, which could be done through the system menu X on the title
bar, later closing of the file browser caused a crash. This is solved by
keeping track of the fact that there is such a daughter through a global
flag, and ignoring the pop-down command if there is one.