Alter treatment of moves with empty squares
In Edit Position mode one can move empty squares, and use them to
'capture' pieces as a means to clear their square. But this often leads
to unintended disappearence of pieces one wants to grab for dragging, if
a previous click had inadvertantly left an empty square selected: the
click on the piece is then seen as the to-click of a click-click move.
  This patch treats to-clicks of moves with an empty-square differently
from other to-clicks: Rather than performing the move (in this case
clearing of the to-square) on the down-click, it defers that to the
up-click, and only does it if that up-click is in the same square. This
allows the user to grab the piece on the down-click, and drag it to
another square, if that is what he wanted to do, without being disturbed
by sudden implosion of the piece he tried to grab.
  If the to-square is in the holdings, grabbing the piece is the only
action, as moves of empty squares into the holdings are forbidden anyway.