Fix switching between pixmap and bitmap pieces
There were still several flags that had to be reset when redoing the
graphics initialization. The Debian patch to use a default bitmap
directory when none is given but monoMode is selected or forced, is
moved to a more sensible place, so that it does not disturb matters when
compiled without LIBXPM (when we do have built-in bitmaps).