Fix chaining of bottom-row dialog buttons
The OK & cancel buttons are chained to the bottom. This patch makes sure
that other buttons appearing in the same row (as in the tags and
comment dialogs) will be chained that way too, so that they will behave
the same on vertical sizing. The bottom of a full-width text-edit above
it will be chained to bottom too. To make this work in the tags dialog,
the (optional) cmail-message field had to be put above the tags memo.