Fix warnings XBoard file browser
All the prototypes have been moved to selfile.h, to force uniformity.
The order of the #include files xstat.h and selfile.h had to be
swapped, to make sure 'struct stat' is a nown type atthe time of
prototype declaration. The event handler SFmotionList is considered
wrong type because of its third argument (XMotionEvent in stead of
XEvent), and had to be casted to (XtEventHandler) when passed to
XtAddEventHandler.