Improve sizing of comment and tags dialogs
There seems to be a bug in Xaw for the sizing of buttons; they get
unpredicatble and weird sizes when the size is not explicitly set,
possibly stretching the initial width of a dialog to rediculous values.
Sizing back to a remembered value from the settings file then makes
buttons overlapand text edits shrink to almost nothing. To prevent that
we now define exlicit sizes for the buttons in tags and comment popups.