Let mentioning completed tourney file add one cycle
Secifying a tourney file of a fully completed tourney will now resume
that tourney with another cycle added. When there are still uncomleted
games in the tourney, the usual error message will appear. There are
also error popups added for when the user fails to specify a tourney
file, or gives too few participants.