Fixed small bugs in several .po files, enabling these translations to be used.  They all probably need more work from the translation team. I don't speak any of these languages, but I did some spot checks either from my own knowledge or using Google Translate, and they generally look sane.
- Several files were in utf8 but needed to be marked as such (that is,
  "CHARSET" changed to "UTF-8").  Affected es.po, it.po, ru.po, vi.po,
  zh_CN.po, zh_TW.po.

- it.po was missing a \n at the end of one translation.

- ru.po had unescaped double-quote marks in one translation.

- vi.po needed to be normalized to Unicode normalization form C.  With
  this fix, more characters display correctly, though at least one is
  still missing in the X core fonts that xboard is able to use.
  See http://vietunicode.sourceforge.net/ for more about Vietnamese.

- zh_HK.po was written in big5 and then utf8-encoded on top of that,
  making it unusable.  I stripped off the spurious utf8 encoding and
  then used recode to convert the file to Unicode properly encoded as
  utf8.

Unfortunately, I wasn't able to test any of the Chinese translations,
as I haven't been able to get XCreateFontSet to work in Chinese
locales with any of the fonts I have installed, even after adding the
Ubuntu language support packages for both simplified and traditional
Chinese.  I did spot check a bunch of the translations using Google
Translate.

I've checked in a couple of tiny Perl scripts that I used to do part
of the fixups on vi.po and zh_HK.po in case we need them again.