Fix concurrency in Swiss tourneys
The pairing engine was consulted for the next round before all games of
the previous round were finished, if other XBoard instances were still
playing games. This caused the instances finishing early to stall
indefinitely. This is fixed by moving the Swiss pairing code to after
the syncInterval code. The normal pairing code (the call to Pairing)
must stay before it, because it determines the syncInterval. So ther is
a bit of tourneyType dependence cluttering up the code here. :-(