Internationalize the file browser.
The file browser was missed in the previous round of
internationalization and had gotten broken by setting international =
True in its widgets.  I got nothing but segfaults trying to set
international to False in the file browser widget subtree when it was
True in the main window, so I gave up and did the
internationalization.  It turned out to be pretty easy.

As a tiny bonus fix, I removed some ../'s from includes, and you can
now configure and build xboard in a subdirectory again.  This was
useful for building both --enable-nls and --disable-nls from the same
sources to compare functionality and make sure both still work.