Make non-existing option in settings file non-fatal
It is rather difficult to recover from a situaton where the settings
file contains an unknown option (e.g. because you downgraded to a
previous version that had fewer options), especially in WinBoard, where
the user settings file is in a hidden folder, and people might not be
able to find it, let alone edit it. By just skipping the line with the
offending option, rather than generating a fatal error, when reading
from a settings file, such options will be automatically purged from the
file as soon as you save settings.
  Small problem is that the user cannot be warned, as the initialization
has not progressed far enough at this point to generate error popups. In
XBoard we can at least print to the console, but in WinBoard there is no
warning at all. But the worst thing that can happen is that new options
in a settings file of a more advanced version will revert to their
defaults.