Cure flicker in Move History window, fix highlighting
The highlighting of the current move in the Move List window by selecting
it did not always work, because some other windows were clearing their text
widgets by selecting the contents and killing it, thus snatching the
selection away. This has been changed to setting XtNstring to an empty
string as a method for clearing without side effects on the selection.
As a bonus the Engine Output window now also no longer has the black flicker
on clearing it.
  In addition, triggering a scroll by setting the insertion point to the
end of the text caused excessive flicker in the Move List window, apparently
clearing it and redrawing from scratch, (even if there was no scroll motion),
to the point where on my slow laptop the move history stayed entirely blank
during a rapid succession of moves. This has been combatted by using the
end-of-line action procedure (which does not seem to suffer from this) to put
the insertion point at the end, when we need to scroll to close to the end.