Add tab stops in WB generic popup
The WinBoard generic popup did not put tab stops on push buttons,
including the OK and cancel button, with as a consequence that these were
skipped when tabbing through the dialog controls. In addition the tabbing
could not begin before the user selected one of the controls with the mouse.
  Now all buttons have tab stops, and the cancel button is given focus
when the dialog pops up.