Fix JAWS reading of Engine Settings dialog
Group boxes are suppressed, as JAWS turns out to be allergic to them
(always reading the entire content of the group box when an item in it
is selected). Put 'browse' on the browse buttons, even though it does
not fit (JAWS reads it anyway). All conditionalon the JAWS flag.