Give WinBoard Game List its own font
The font fort he Game List used to be shared with the Engine Output and
Move History windows, but this made it impossible to use a figurine font
in the latter two (which displays piece symbols for capitals) without
messing up the Game List. An option -gameListFont has been added to specify
the font for the Game List(and save it in the settings file), while a
control for it is added in the fonts dialog.