Force Move History refresh after loading/reverting variation
The Move History window was not refreshed after switching to or from
a variation if that variation happened to have the same length as
the main line. Now a change in storedGames since the previous refresh will
force a full refresh to avoid this. For this to work, storedGames
did have to be decremented after ToNrEvent in PopInner, because ToNrEvent
calls HistorySet before the history is restored, so we don't want
it to snatch away the full refresh. A call to HistorySet had to be
added after PopInner is done restoring the move history.