Allow entry of variations in PlayFromGameFile mode
Adding variations is more like editing a comment,and not really editing
the game. Therefore we allow it on loaded games. As soon as we are in a
variation, we temporarily switch to a (hidden) EditGame mode, restoring
the original mode when we revert to the main line.
  We now also stay in PlayFromGameFile mode after loading a game,
And opening variation (fromtheComment window) in PlayFromGameFile mode
is also allowed.