Implement searching games in Game List for a position
For this it was needed to reconstruct all positions of the games
without loading them, so they could be compared to the currently
displayed position. A new routine GameContainsPosition is made for that.
 In Disambiguate, when the global quickFlag is set, we now do the
disambiguation ignoring check,and only if this leads to an ambiguous
move, we repeat it taking check into account to see if that resolves
the ambiguity. This causes us to miss if unambiguous moves are illegal
because they put their King in check, but who cares?
 We now do not make game-list line when not needed. If we do not filter
by text, but by position, there is no need for preparing the game-list
line for games that are not selected. All for efficiency.