Allow deletion of last book move
The book-edit function refused to delete the entiremove list,
for no good reason. This limitation now has been lifted.
Also the error message that the key for the current book entries
is different from that of the currentposition has been suppressed
if there were no entries found for the current position (so that the
book is positioned at the next key), so that it is possible to add
moves from scratch.