Add key bindings for loading next/previous game of list
These bindings are set both for the game list and the board window.
To make them useful, focus has to be transferred to the board widget
after loading (it tends to transfer to the tags window), which now
seems to work, but not always. From the game list the shellWidget
seems to get focus rather than the board, which is bad as it is not
sensitive to the key bindings.