Allow substitution of engines during tournament
Two buttons are added in the tournament options dialog, for upgrading
and for replacing a tourney participant. Lots of tests on the validity
of the request are done, and if all are passed, the tourney file is
written with the new participants (and in case of replace) with the
results of the replaced engines erased from the -results string.