Add -useBoardTexture and -usePieceFont options
These options suppress the use of the given texture files or piece fonts
without relying on the kludge to disable them by prefixing an asterisk
to the file or font name. Checkboxes have been added to the Board dialog
to control them. The user can now set piece font in WB font dialog.
The sample text can be used to specify the -fontPieceToCharTable,
the font name itself the -renderPiecesWithFont value.