Allow two-games-per-opening to work with book
A new option -seedBase is added, and written to the tourney file.
It allows all XBoard instances working on a tourney to generate the same
'random' openings in a deterministic way, based on this one-time seed
and the game number. When -loadGameIndex equals -2, it uses the same seed
for an odd and the following even game.