Implement paging in XBoard Game List
The XBoard Game List window could not display more than ~1900 games,
probably due to limitations to an Xaw listbox. Now the games are listed
in 'pages' of 1000 games, through which one can step by clicking the
"previous page" and "next page" lines at the top or bottom of the listbox.