Speed up position search and consider side to move
The position search is made to pay attention to the side to move,
which produces a speedup, because we only have to compare half the
game positions when looking for an exact position match. An addition
we now keep track of the total number of pieces, and abandon a game when
it drops below the number of pieces in the position we seek.