Auto-popup ICS Input Box
Typing to the board made the move type-in pop up, but in ICS mode
this is of course undesirable. Like in WinBoard, the ICS Input Box
should pop up in stead, and accept the typed character. This is now
realized. If the box already exists and has text in it, the typed
character is appended to that text, and the box is given focus.
Spaces now also are considered as printable.