Upgrade -serverMoves option
The (undocumented) -serverMoves option was upgraded to conform to the
ChessLive! protocol extension for sending SAN tags. A bug that allowed
upper-case promoChar to leak through was also fixed, and code was added
to write engine PVs to file. Flushing of the game result to the -serverMoves
file was also fixed.