Make too-long game non-fatal
GameEnds is now called when the maximum number of moves is exceeded,
rather than DisplayFatalError. Not sure if this properly terminates the
game in all situations, but DisplayFatalError (after calling GameEnds)
would also wait for the user to confirm the error popup before really
exiting, so that in an unattended automatically running tourney there
was not much fatal, and the entire tourney would in fact be finished.