Limit width of menu bar in XBoard
This is a rather course hack to prevent that the menu bar well determine
the window width, leaving white space next to the board, for small
square size. Rather than extending the tinyLayout regime, where only the
first letter of the menu names would be printed to save space, this
patch causes a gradual transition, where the number of letters shown
for each menu item goes down with the board width.