Reactivate Falcon Chess
Now the patent for this has expired, Fairy-Max plays it again.
So the XBoard support for it is now switched on, and an item for it is
added to the XBoard New Variant menu dialog. (WinBoard still had that.)
In the array the Lance was changed for the Falcon as a wild-card piece.