Add option -discourageOwnBooks
This boolean option determines the default setting of the hasOwnBook
option before a tourney game starts. By setting it, the hasOwnBook
option is cleared, which would enable use of the GUI book, like the
engine was installedwith the /firstXBook option. An explicit option
on the engine install line can still overrule this, though.
 A control to set this option has been added in the XBoard Match dialog,
and it is saved in the tourney file to force consistency.