Fix duplicate loading of second engine in tourneys
To decide if new engines should be loaded, it was tested if first.pr
equalled NoProc, assuming that ReserveGame would have left the engine
loaded if the next game would need the same engines. But a reuse=0 engine
could have caused termination of the first engine process anyway, while
the second would still be alive. So unconditionally starting of the second
engine would cause redundant engine processes to abound.
  Now we only start new engine processes if both engine processes are dead.