Add mechanism to translate variable messages in WinBoard
A message prefixed with % in the lng file will be considered a match
if it matches the start of the text to be printed, and the remaining
part of the latter will remain untranslated. This was needed to allow
translation of the perpetual change message, now it indicates the square
of the chased piece.