Clear board markers when moving to other position
Now the markers from -showTargetSquares would stay on after a first click
for the benefit of a click-click move, they would also stay on when not
finishing the move, but operating the < > buttons in stead to call up
other positions. ForwardInner and BackwardInner now clear the markers,
like they also cleared highlights.