Fix some ICS issues and remove the Xt code from xoptions.c
Some buttons on the ICS text menu were not working properly (those
with a name parameter e.g. finger). These should now be fixed so you
can select some text in the terminal window with the mouse and then
click a button and it will use the selected text in the command.