Add feature-override options
A few (dangerous) options are added to allow a user to override
engine features, or change their defaults, by supplying a string
with feature commands (e.g. "sigint=0") which tricks XBoard into
thinking the engine sent that. There is a -featureDefaults which is
processed before the engine features, and -first/secondFeatures
which is parsed just before engine initialization.
  This goes horribly wrong when the user wouldtry silly things like
including 'option' or 'done' features!