Fix WinBoard  clock-click bug
Clicking on the clock in WinBoard also called the LeftClick routine
for handling clicks on the board. This led to the from-square being set
to some inalid value, so that the next genuine board click was interpreted
as a to-click, with undesired effects, especially with legality testing off
or in EditPosition mode, where weird things are accepted.