Add NO_GETTEXT flag for combobox. Fix bug #34991
Define NO_GETTEXT flag to avoid calling gettext() for comboboxes, and define COMBO_CALLBACK
flag to call comboCallback() when an item is selected in a combobox (no new code for this,
just a symbolic value instead of a number). The flags are used for the min field of struct Option.
NO_GETTEXT is used for the Select engine combobixes in the Match Options and Load engine windows.