Make GenericPopUp() more readable by using more named flags instead of numerals
Defines the flags SAME_ROW (value 1) and NO_OK (value 2) for use with buttons and endmarks.
There should no functional change.
Also add a code comment where a condition depends on an uninitialised value. There is no direct harm, as both
branches do the same (that is nothing) when the value is uninitialised.