Fix possible crash on 64-bit systems when copying game or position
The two convertproc's for X selections caused reads out of bounds on 64-bit systems.
I hope that this is the right solution. I am not experienced with X stuff, but it
follows that I have seen in other programs, testing shows that copying still works,
and the reading out of bounds have disappearinged with this patch.