Implement -viewerOptions
A mechanism is added to allow the user to configure persistently if the
game viewer that starts on clicking PGN files will use -ncp mode, or
load an egine for analysis (or, in fact, do anything). This by introducing
a persistent string option -viewerOptions, which will be parsed as a
command line after the real command line when the option -viewer was present.