Update texi file and remove duplicate control
Some missing option descriptions were added. The highlight-dragging
control was removed from the General Options dialog, as it did the same
as Show Target Squares. -highlightDragging wa changed to a volatile option
with default value true, making it effectively disappear as an XBoard
option. (In WinBoard it is still persistent.)