Fix seek-graph popup and  popdown
Clicking outside the board area in WB (e.g.on the clocks) would also
trigger seek-graph 'popup'. This is now suppressed by testing for
y-coord > 0. In addition, the seek graph is automaticaly taken down
when we step to another position in the current game (using buttons,
arrow keys or whatever).