Prevent unnecessary loading of engines
For engines that can be reused, the unloading at the end of a pairing
is in some cases wasteful, as the same engine will be used in the next
pairing as well. In simple gauntlets the gauntlet engine will always
stay first engine, while in multi-gauntlets the oppent engine will stay
the same for an entire round. In these cases unloading is now suppressed.