Fix arrow damage with highlightDragging XB
With highlight dragging the highlights of the previous moves are cleared
when lifting the piece. Logically, this should also clear the arrow, but
to do that requires a board redraw, which is not done at this stage. (And
even then this would only be effective after marking the arrow damage,
which is not done yet either). As a temporary fix, we repair the arrow
by redrawing it.
  This patch has moved marking of the arrow damage from DrawArrowHighlight
to a separate routine, so a future better fix can also call that when the
arrow should be erased.