Configure XBoard engines and textures
The new -useBoardTexture option is needed in xboard.conf, as we want to
distribute with the texture on, while the compiled-in default is off.
A list of 4 engines has been configured (fairymax, fruit, crafty, gnuchess),
as the Load Engine and Match Options checkboxes crash XBoard when the list
is empty. The -recentEngines option is set to 0 to make sure the code
associated with it is never executed: the code is still experimental,
and the required front-end support is only present in WinBoard anyway.