Better solution to button-height problem
In stead of adapting the height of text widgets to that of buttons,
(as represented by the message widget), we now do the opposite, and
adapt the button & label heights. This is made safe by obtaining textHeight
from the fontSet's max_logical_extent, though an XExtendsFontSet() call.
Turns out the default height of text widgets is not exactly that, but
4 pixels larger for borders and margin.
  The parameter XtNvertDistance is now used to align the checkboxes with
the following text widget, and their size is derived from textHeight as
well.
  The XtNvertDistance is set to -1 for the lower spin button, so that it
its border coincides with that of the upper spin button. Their height is
set to textHeight/2. This makes their combined height just 1 pixel more
than that of the preceeding text widget. (Uh? 2 x tw/2 -1 > tw?), so for
other buttons we use that too (i.e. textHeight = font_extent + 5).